Renommage et mutualisation du code
