+++ front/gatsby/src/components/Citation/Citation.module.scss
@@ -1,0 +1,95 @@
+@use '../../styles/defaults.scss' as *;
+@use '../../styles/variables.scss' as *;
+
+.citation {
+  align-items: center;
+  background-color: $extra-background-color;
+
+  &:nth-child(2n + 1) {
+    background-color: $main-background-color;
+  }
+}
+
+.citationTypeIcon {
+  width: 1em;
+  height: 1em;
+  margin-right: 0.35em;
+}
+
+.citationName {
+  overflow-x: hidden;
+  text-overflow: ellipsis;
+}
+
+.copyToClipboard {
+  margin-left: auto;
+}
+
+.responsiveTable {
+  margin-top: 0.5em;
+  margin-bottom: 0.5em;
+}
+
+.citationList {
+  width: 100%;
+  border-collapse: collapse;
+
+  td,
+  th {
+    padding: 5px;
+  }
+}
+
+.colIcon {
+  width: 5%;
+
+  img {
+    width: 1em;
+    height: 1em;
+    margin: auto;
+    display: block;
+  }
+}
+
+.colKey {
+  text-align: left;
+  word-wrap: break-word;
+  max-width: 1px;
+  font-size: 0.9em;
+}
+
+.colActions {
+  text-align: right;
+  width: 10%;
+}
+
+.more {
+  font-size: 0.9rem;
+  text-align: center;
+  color: #727272;
+}
+
+.searchField {
+  input {
+    width: 100%;
+  }
+
+  padding: 1rem 0 0;
+}
+
+
+.citations {
+  > textarea {
+    width: calc(100%);
+    padding: 1rem;
+    height: 5rem;
+    margin-bottom: 1rem;
+  }
+}
+
+.citationMessages {
+  font-size: 0.8rem;
+  margin: 0.5em 0 1em 0;
+  color: $error-color;
+}
+
