+++ front/gatsby/src/components/Citation/CitationTypeIcon.module.scss
