+++ front/gatsby/src/components/Citation/ReferenceList.module.scss
@@ -12,10 +12,3 @@
   padding: 0 0.5rem 1rem 0;
   justify-content: flex-end;
 }
-
-.more {
-  font-size: 0.9rem;
-  display: flex;
-  padding-bottom: 1rem;
-  justify-content: center;
-}
