--- front/gatsby/src/components/ReferenceTypeIcon.jsx
@@ -1,13 +1,0 @@
-import React from 'react'
-
-import { iconName } from '../helpers/bibtex'
-import styles from './ReferenceTypeIcon.module.scss'
-
-export default ({ type, className }) => (
-  <img
-    src={`/images/bibtex/${iconName(type)}.svg`}
-    className={[styles.icon, className].join(' ')}
-    alt={type}
-    title={type}
-  />
-)
