+++ front/gatsby/src/components/Write/bibliographe/bibliographe.module.scss
@@ -34,7 +34,6 @@
 
 
 .raw,
-.citations,
 .zotero {
   margin-bottom: 1rem;
 }
@@ -63,77 +62,7 @@
   }
 }
 
-.citations {
-  > textarea {
-    width: calc(100%);
-    padding: 1rem;
-    height: 5rem;
-    margin-bottom: 1rem;
-  }
-}
-
-.responsiveTable {
-  max-height: 270px;
-  overflow: auto;
-  margin-top: 0.5em;
-  margin-bottom: 0.5em;
-  border-top: 1px solid #bbbbbb;
-  padding-top: 0.25em;
-}
-
-.citationList {
-  width: 100%;
-  border-collapse: collapse;
-
-  td,
-  th {
-    padding: 5px;
-  }
-}
-
-.colIcon {
-  width: 5%;
-
-  img {
-    width: 1em;
-    height: 1em;
-    margin: auto;
-    display: block;
-  }
-}
-
-.colKey {
-  text-align: left;
-  word-wrap: break-word;
-  max-width: 1px;
-  font-size: 0.9em;
-}
-
-.colActions {
-  text-align: right;
-  width: 10%;
-}
-
 > p.citation {
   padding: 0.5rem 1rem;
 }
 
-.citationMessages {
-  font-size: 0.8rem;
-  margin: 0.5em 0 1em 0;
-  color: $error-color;
-}
-
-.searchField {
-  input {
-    width: 100%;
-  }
-
-  padding: 1rem 0 0;
-}
-
-.resultFoundCount {
-  font-size: 0.9rem;
-  display: flex;
-  justify-content: flex-end;
-}
