--- front/gatsby/src/components/Write/reference.module.scss
@@ -1,28 +1,0 @@
-@use '../../styles/defaults' as *;
-@use '../../styles/variables' as *;
-
-.reference {
-  display: flex;
-  align-items: center;
-  background-color: $extra-background-color;
-
-  &:nth-child(2n + 1) {
-    background-color: $main-background-color;
-  }
-
-}
-
-.referenceTypeIcon {
-  width: 1em;
-  height: 1em;
-  margin-right: 0.35em;
-}
-
-.referenceName {
-  overflow-x: hidden;
-  text-overflow: ellipsis;
-}
-
-.copyToClipboard {
-  margin-left: auto;
-}
